requirements.txt
pymongo==4.5.0
psycopg2-binary==2.9.9
Faker==19.13.0
